Fetch transaction details as well & refactoring
In the process I renamned the previous Transaction to Transaction.Position and
the details now live as transaction with an associated list of positions. Makes
more sense that way.

I kinda dislike that both requests are currently sent of after each other, but
didn't want to mess around with GCD right off the bat. Will hopefully come back
to fixing that.

Besides that this actually seems to be working rather well. Didn't expect that 😄